Move ICA driver methods into their own module as functions
